# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "APi de Personajes y Pelicualas/Serie 3.0",
    "description": "Este es un API de Peliculas/Series y Personajes los cuales estan asociados entre si",
    "contact": {
      "email": "ulibaamonde@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
...
  "paths": {
...
    "/characters": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "Busqueda por parametros",
        "description": "Mediante este endpoint, podes conseguir el ID, Nombre y Imagen de un personaje de la Api con los siguientes parametros",
        "operationId": "getByParams",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "El nombre del personaje",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "age",
            "in": "query",
            "description": "La edad del personaje",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "movie",
            "in": "query",
            "description": "El edad de la pelicula relacionada al personaje",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personaje/s encontrado/s",
            "headers": {
              "X-Rate-Limit": {
                "description": "calls per hour allowed by the user",
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "X-Expires-After": {
                "description": "date in UTC when token expires",
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No se encontro ningun Personaje"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "characters"
        ],
        "summary": "Creacion de Personajes",
        "description": "Con este endpoint podes crear personajes para agregar a la Base de Datos",
        "operationId": "createCharacters",
        "requestBody": {
          "description": "Ingrese todos los necesario para crear el personaje",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Personaje"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Personaje"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Personaje"
              }
            },
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Personaje"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Personaje"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Ingreso de datos incorrecto"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "characters"
        ],
        "summary": "Actualizar Personaje",
        "description": "Con este endpoint se puede actualizar los datos de un personaje ya existente en la base de datos",
        "operationId": "updateCharacter",
        "requestBody": {
          "description": "Ingrese los datos necesarios para hacer el remplazo",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Personaje"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Personaje"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Personaje"
              }
            },
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Personaje Actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Personaje"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Personaje"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Ingreso de datos incorrecto"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/characters/{id}": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "Busqueda por ID",
        "description": "Con este endpoint podes buscar toda la informacion de un personaje usando su Id",
        "operationId": "getByID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id usado para buscarlo",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personaje encontrado"
          },
          "400": {
            "description": "Id invalido"
          },
          "404": {
            "description": "Ningun Personaje encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "characters"
        ],
        "summary": "Borrar Personaje",
        "description": "Con este endpoint podes borrar un personaje usando su ID",
        "operationId": "deleteByID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id usado para buscarlo",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personaje borrado"
          },
          "404": {
            "description": "Personaje no encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
...
      "Personaje": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int64"
          },
          "Imagen": {
            "type": "string"
          },
          "Nombre": {
            "type": "string"
          },
          "Edad": {
            "type": "integer",
            "format": "int64"
          },
          "Peso": {
            "type": "integer",
            "format": "int64"
          },
          "Historia": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Personaje"
        }
      },
...
    },
...
  }
}